packages/em-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "em-core"
version = "0.1.0"
description = "Executable Memory — convert agent traces into deterministic routines"
readme = "../../README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Sami Veikko"}]
keywords = ["agent", "trace", "routine", "deterministic", "automation", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0,<3",
    "typer>=0.9,<1",
    "pyyaml>=6.0,<7",
    "jinja2>=3.1,<4",
    "jsonschema>=4.0,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
em = "em.cli:app"

[project.urls]
Homepage = "https://github.com/samiveikko/executable-memory"
Repository = "https://github.com/samiveikko/executable-memory"
Documentation = "https://github.com/samiveikko/executable-memory/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["em"]

[tool.pytest.ini_options]
testpaths = ["tests"]
